+++ src/BooksApp.Infrastructure/Bookshelves/Persistence/BookshelvesConfiguration.cs
@@ -1,7 +1,7 @@
 ﻿using BooksApp.Domain.Book;
 using BooksApp.Domain.Bookshelf;
 using BooksApp.Domain.Bookshelf.ValueObjects;
-using BooksApp.Domain.Common.Enums.MaxLengths;
+using BooksApp.Domain.Common.Constants.MaxLengths;
 using BooksApp.Domain.User.ValueObjects;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
@@ -21,7 +21,7 @@ public class BookshelvesConfiguration : IEntityTypeConfiguration<Bookshelf>
                 value => BookshelfId.CreateBookshelfId(value));
 
         builder.Property(g => g.Name)
-            .HasMaxLength((int)BookshelfMaxLengths.Name)
+            .HasMaxLength(BookshelfMaxLengths.Name)
             .IsRequired();
 
         builder.Property(b => b.UserId)
